feat: introduce decimal score, get more accurate cutle score (2 decim…

+++ backend/Models/Submission.cs
@@ -21,8 +21,8 @@ namespace DailyChallenges.Models
         [Required]
         public string Score { get; set; } = string.Empty;
 
-        // Numeric score value parsed when the score can be represented as an integer
-        public int? ScoreValue { get; set; }
+        // Numeric score value parsed from the Score string (supports decimal values)
+        public double? ScoreValue { get; set; }
 
         public string? Username { get; set; }
